cleanup

+++ recipe/build.sh
@@ -3,9 +3,11 @@
 set -ex
 
 ./configure --prefix=$PREFIX
-cat config.log
-# avoid libc++ conflict with <version>
-rm -rf VERSION
+
+if [[ "$target_platform" == osx-* ]]; then
+  # avoid libc++ conflict with <version>
+  rm -rf VERSION
+fi
 
 make -j${CPU_COUNT}
 make test
